requirements.txt
pip
flask
flask_restplus
tensorflow
keras
numpy
Pillow
werkzeug
setuptools
virtualenv
firebase_admin
pandas
